Fix gameover.gr articles
The main text in each article is currently being hidden if this list is enabled, and cosmetic filtering is also enabled. The body of each article is inside a `share-container`.

I hope this is the correct way of fixing it. I'm not familiar with rules and I don't see any rule blocking `share-container`s, so I assume the `~rockinathens.gr##.share-container` line blocks any `share-container` but the one from rockinathens. If that is the case perhaps it would be better to remove the rockinathens.gr rule instead, since it is likely that other pages will use the same layout with the main text inside a `share-container` element.